Refactoring CSS to its dedicated folder
