+++ src/css/Acties.scss
@@ -1,3 +1,5 @@
+@import "variables";
+
 #acties {
     padding: 16px;
 
@@ -6,7 +8,7 @@
     }
 
     .points{
-        color: blue;
+        color: $blue;
         font-weight: bold;
         margin-top: 0;
         font-size: 18px;
@@ -17,7 +19,7 @@
         width: 100%;
         padding: 16px;
         box-sizing: border-box;
-        background-color: #EEEEEE;
+        background-color: white;
         border-radius: 5px;
 
         box-shadow: 0 1px 3px rgba(0,0,0,0.12), 0 1px 2px rgba(0,0,0,0.24);
