+++ src/css/Alert.scss
@@ -1,3 +1,5 @@
+@import "variables";
+
 #alert{
     color: red;
     text-align: center;
