+++ src/css/Login.scss
@@ -1,6 +1,7 @@
+@import "variables";
+
 #login, #register {
     padding: 15px;
-
     .logoImg {
         margin: 0 auto;
         width: 60%;
@@ -21,7 +22,7 @@
         font-size: 1.2rem;
         border-radius: 0.5em;
         border: none;
-        box-shadow: 1px 2px 4px rgba(0, 0, 0, .5);
+        box-shadow: $boxshadow;
     }
 
     ::placeholder {
@@ -40,13 +41,13 @@
         padding: 15px;
         min-width: 30%;
         color: white;
-        background-color: #00337f;
+        background-color: $blue;
         border-radius: 0.5rem;
         border: none;
-        box-shadow: 1px 2px 4px rgba(0, 0, 0, .5);
+        box-shadow: $boxshadow;
 
         &:hover {
-            background-color: #009cde;
+            background-color: $hoverblue;
         cursor: pointer;
         }
     }
@@ -55,11 +56,11 @@
         display: block;
         padding: 15px;
         font-size: 1.3rem;
-        color: #00337f;
+        color: $blue;
         border:none;
 
         &:hover {
-            color: #009cde;
+            color: $hoverblue;
         }
 
     }
