+++ src/css/_variables.scss
@@ -2,3 +2,5 @@
 $orange: #FFC61E;
 $backgroundOrange: #FFF8E2;
 $blue: #003373;
+$hoverblue: #074973;
+$boxshadow: 1px 2px 4px rgba(0, 0, 0, .5);
